code review updates
