+++ e2e_example/tool/build.dart
@@ -1,6 +1,8 @@
 // Copyright (c) 2016, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:io';
+
 import 'package:build/build.dart';
 
 import 'package:e2e_example/copy_builder.dart';
@@ -16,5 +18,9 @@ main() async {
 
   var result = await build([phases]);
 
-  print(result);
+  if (result.status == BuildStatus.Success) {
+    stdout.writeln(result);
+  } else {
+    stderr.writeln(result);
+  }
 }
