+++ e2e_example/tool/watch.dart
@@ -1,6 +1,8 @@
 // Copyright (c) 2016, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:io';
+
 import 'package:build/build.dart';
 
 import 'package:e2e_example/copy_builder.dart';
@@ -15,6 +17,10 @@ main() async {
   var phases = CopyBuilder.buildPhases(graph);
 
   await for (var result in watch([phases])) {
-    print(result);
+    if (result.status == BuildStatus.Success) {
+      stdout.writeln(result);
+    } else {
+      stderr.writeln(result);
+    }
   }
 }
